requirements.txt
branca==0.6.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.0.1
click==8.0.1
colorama==0.4.6
cryptography==37.0.2
Flask==2.0.1
folium==0.14.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.24.2
peewee==3.14.10
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.17.1
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.0.1
